+++ frontend/src/App.js
@@ -20,7 +20,6 @@ import ClientFormPage from "./pages/ClientForm/ClientFormPage";
 import CompanyFormPage from "./pages/CompanyForm/CompanyFormPage";
 import CompanyPage from "./pages/Company/CompanyPage";
 import CompanyDetails from "./pages/CompanyDetails/CompanyDetails";
-import TestPage from "./pages/Test/TestPage";
 import ExpensePage from "./pages/Expense/ExpenseFormPage";
 
 const App = () => {
@@ -53,7 +52,6 @@ const App = () => {
                 <Route path="/company-form" element={<CompanyFormPage />} />
                 <Route path="/company-details" element={<CompanyDetails />} />
                 <Route path="/expense" element={<ExpensePage />} />
-                <Route path="/test" element={<TestPage />} />
             </Routes>
         </Router>
     );