+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+-   Add employee name to report PDF
+-   Add company data and logo to report PDF
+-   Add dynamic title page on mobile
+-   Add Dockerfile for frontend/backend and add a docker-compose file to run the app on Docker
+-   Add client details page to mobile
+-   Update add company forms
+-   Update CD pipeline to upload a updated images for each release
+
+# [0.3.4] (2024-12-03)
+
 -   Update report clients and report page on mobile
 -   Add links to mobile menu